Auto mime-types. Lookup tool.
We now also leave the timer objects to be autostarted by our timer-management.
This can be globally toggled to keep from spawning additional threads when not
appropriate. We've also added a tool to dump entries based on ID or file-path.
We now try to derive a mime-type for a new file, if not provided.