Fixed binaries. Fixed deps. Up'd README.
The binaries weren't being included in PyPI upload. They have to have a ".py" extension. We removed the Google client as a dependency because it broke our PyPI install EVERY TIME. We updated the README to encapsulate all of the negative phenomena that we've experienced, as well as to provide specific install steps for pip, easy_install, manual, etc..